mononoke: expose PerfCounters at request level
Summary: One thing we have right now in Repo Client is logging of perf counters (this happens in the call to finalize_command). However, we don't have this at the request level. This is a little annoying because there is some context about the client calling us that only exists in the request level, so we can't aggregate on PerfCounters there.

Reviewed By: krallin

Differential Revision: D26257247

fbshipit-source-id: 220c1adafd583420e64599befbd165152c3a8c6f