mononoke/edenapi_service: check & bump load counters for trees & files
Summary:
Like it says in the title. Let's add some basic throttling control here, in
line with what we have in Mononoke Server. The numbers don't quite match up
since fetches in EdenAPI don't include linknodes or history, but this should be
better than nothing and sufficient for now, and makes sense to have with
EdenAPI & Mononoke Server running in the same process.

Reviewed By: HarveyHunt

Differential Revision: D26250746

fbshipit-source-id: 338eda4341a163d0d915f10bf45fc7f40c74fc69