mononoke: add load limiter to EdenAPI sessions
Summary:
Like it says in the title. I'd like to access our load limiter from there and
this is a good way to do it.

Note that EdenAPI doesn't have client hostnames when we instantiate the load
limiter, but we also don't truly need it because our client identities include
the hostprefix (at least in cases where a reverse lookup stands a chance of
succeeding, i.e. in prod).

Reviewed By: HarveyHunt

Differential Revision: D26250743

fbshipit-source-id: d1ad062d2967bfe28b3ddc164cac977d297d1b6a