mononoke/hooks: extend file fetcher to find files
Summary:
Owners file hook requires access to the files in master branch.
This diff extends `FileContentFetcher` to have API for finding files in the master commit. The methods are implemented for the blobrepo. They derive unode manifest for the master bookmark and search given file paths in the manifest, returning list of file content ids.

Reviewed By: StanislavGlebik

Differential Revision: D24936674

fbshipit-source-id: c63931c073b6148972bc8e81126f7c4786db920a