mononoke/redaction: remove convoluted & useless timestamp check
Summary:
So right now we get the last timestamp at which we logged, then do a compare
exchange and log only if that succeeds. So we're ... racing with another
thread being on the brink of logging too?

What is that possibly supposed to achieve? Let's remove it.

While in there, also clean up the imports a bit, fix up the i64 -> u64
conversion, and remove a useless borrow.

Reviewed By: ahornby

Differential Revision: D26337990

fbshipit-source-id: a17f6613f1fb5fa58e4131181c230006af06f003