mononoke/hooks: resolve owners rule
Summary:
This diff adds a function that for a given path and a parsed path tree of owners files finds a rule that is going to be applied for the path to check permissions.

Because of how the owners files work, we care about the closest owners file rule to the path. So in the implementation I go from the bottom to up in the owners files path tree.

Reviewed By: StanislavGlebik

Differential Revision: D25379330

fbshipit-source-id: 43b95938c90fdd2e1ef4fe80ed400d171e1df459