fs: default initialize XdrVariant
Summary:
My understanding of C++ not being perfect, I thought that these 2 fields would
be default initialized, but it turns out this is wrong, and the values of these
would be whatever is on the stack at the time a non-initialized variant is
used. The XdrSerialize variant test exercise this case and fails when optimized
due to that, this change fixes it.

Reviewed By: chadaustin, fanzeyi

Differential Revision: D26371695

fbshipit-source-id: 489cac846b51508809783bb577c7a97c4e6c060a