fscap: EdenFS on macOS may be case insensitive
Summary:
On macOS, EdenFS can be configured to be case insensitive, and will soon be
switching its default to be case insensitive. Mercurial however considers that
EdenFS is always case sensistive on unix (linux, macOS), so let's change it on
macOS so it manually checks if the working copy is.

Long term, EdenFS on macOS will always be case insensitive, at which point this
code will be changed again to indicate that and avoid the manual checks.

Reviewed By: chadaustin

Differential Revision: D26357816

fbshipit-source-id: 4022c097f2804da69cfcc176840683a6dca12ffd