mononoke: disable tokio coop
Summary:
See the earlier diff for what flag controls.

When booting SCS, we poll a lot of nested FuturesUnordered. This results in
very inefficient behavior in Tokio's cooperative scheduling, and results in us
spending 50% of total our CPU (in fact, a full thread) on just yielding, with
most other threads being idle.

With this change, we use 20+ threads running work that is scheduled by the main
thread, which is what we want.

Note that this applies to all Mononoke binaries. This has only been especially
bad in SCS startup, but we've also not benefited from this feature anywhere,
so rather than leaving this footgun in other apps, let's take it out
everywhere.

Reviewed By: markbt, StanislavGlebik

Differential Revision: D26399889

fbshipit-source-id: 0a13e1275d367e49c2342cb85cb6cd0047cda224