metagit: add faster repository validaton
Summary:
For some repositories (e.g. configerator-sitevars) running "git gc" is too
slow. At the same time after initial rsync it would be good to run at least
some kind of validation to check that repo is not completely corrupt.

This diff adds that. This code checks that HEAD and 100 of its ancestors have
the object that they reference are reachable.

Reviewed By: krallin

Differential Revision: D26423769

fbshipit-source-id: a3fcd9fc5c30e5bf0fdc1cd0fb9e03bdc2e1371d