mononoke: update walker to use new bounded_traversal_unique
Summary:
The walker mostly checks for duplicates before emitting a new edge, at the same time recording the edge as visited to prevent duplicate edges.

However for derived data where the node may or may not be present, the node isn't considered visited until the node data is successfully loaded and seen in state.rs record_resolved_visit().  This leaves a gap where we could be executing multiple copies of the same node.

Reviewed By: farnz

Differential Revision: D26319139

fbshipit-source-id: 52ce28f15341f132d94ebc1ff5e8ee2f0dc2564a