mononoke: add duplicate detecting variant of bounded traversal stream
Summary:
Sometimes many nodes will unfold with a common child, in which case it is desirable that they aren't scheduled at the same time.

This adds bounded_traversal_unique as a new variant to prevent those duplicates.

Doing the duplicate detection inside bounded_traversal_unique means we only need to keep scheduled_max copies of keys and can keep it in a regular HashMap, vs doing it at bounded_traversal_stream call sites which would mean keeping a copy for everything in the unscheduled queue and using DashMap or a lock around HashMap.

Reviewed By: farnz

Differential Revision: D26319137

fbshipit-source-id: 3174ce9e7db4edeb107d26f72575de886e6b2e39