mononoke: edenapi: Fix memory leak caused by RequestContext
Summary:
Over time EdenAPI server anonymous memory usage has been increasing in
line with the number of requests served. This is due to references to
`RequestContext` outliving the lifetime of a request, causing the contained
data to never be dropped.

`RequestContext::new()` spawns a task that is used for logging errors. The task
uses a clone of `RequestContext` and then loops over the messages received on
`error_rx`. However, the only way for `error_rx.next()` to return `None` is for
`error_tx` to be dropped which happens when `RequestContext` is dropped. As the
task has cloned `RequestContext`, `error_tx` won't be dropped whilst the task
is running. As a result, `RequestContext` won't get dropped until the server
shuts down.

Update `new()` to only clone the logger rather than an entire `RequestContext`.

Reviewed By: mzr

Differential Revision: D26451527

fbshipit-source-id: 535ed53d76830df7a5c98694b4293260b7d677cf