metagit: make fast_validate_repository test faster
Summary:
krallin noted that this test takes 60 seconds to finish. This happens because
of r2d2 connection timeout - by default it's 30 seconds, and
fast_validate_repository test fails while we are trying to get git connection.
Instead of failing immediately it waits for 30 seconds (I looked into r2e2
implementation - when trying to get a connection it waits on a condition variable which never gets
notified because of a failure).

to mitigate the issue let's return Result<..., Error> now

Reviewed By: krallin

Differential Revision: D26462872

fbshipit-source-id: 93bb38da20459dca4e3737352ab3b638e8e88a84