admin: create a subcommand for a rebase
Summary:
This allows to create a changeset clone with p1 changed. Essentially, this is a
bonsai-only rebase without any safety checks. The safety of this operation
(i.e. the usability of the resulting changeset) is highly dependent on the
context the ultimate responsibility is on the person running this command. For
instance, this does not do *any conflict checks*. As such, this is really only
useful to rebase additions/deletions/modifications of unchanged files.

Reviewed By: StanislavGlebik

Differential Revision: D25947017

fbshipit-source-id: 4093c432e4f8ba2a647ee4b17c6840c33d13be81