Update generations in sqlblob on non-overwrite
Summary:
We rely on puts to the blobstore setting the generation number on chunks to ensure that there cannot be a race between puts, marks and deletes.

While not doing so should be harmless assuming mark is bug-free, we should still update the generation number on all referenced chunks when we elide a write, so that mark knows to ignore this chunk for now.

Reviewed By: mitrandir77

Differential Revision: D26469954

fbshipit-source-id: 32decedc69c8d4dcb1071a622b28e98ee36ea238