mononoke/hooks: extend file content manager to search for last updates
Summary:
OWNERS hook sometimes needs to know the most recent changes made to the OWNERS files. So the user, who was last to change the OWNERS file, can change it again.

This diffs extends `FileContentManager` trait to have a new method `recent_changes`. It derives unode manifest for the given bookmark, finds given paths and gets linknodes. For the linknodes it derives `ChangesetInfo` which is a lightweight version of the bonsai changesets, they contain all the metadata, but don't have a set file changes.

We use unodes here instead of hg manifests, because hg manifests don't have correct linknodes.

Reviewed By: StanislavGlebik

Differential Revision: D26468603

fbshipit-source-id: 78dd0a4e51bee6228de1296eafe7ffa393df2fbe