eden/mononoke/commit_rewriting/backsyncer: remove tokio 0.1 spawns
Summary:
This is using `spawn_future` (aka tokio 0.1 spawn) so it needs changing.
Amusigly, this uses it to spawn a 0.2 future and then it has to pull 0.1
futures to poll it.

Well, anyhow, this fixes this, this removes all the manual instantiation
of a compat runtime (because this never needed one), and it also removes 0.1
futures (because we never needed them).

Reviewed By: StanislavGlebik

Differential Revision: D26485584

fbshipit-source-id: f49c85d2a1dc406e2ad9e7998d9270fe2e6c9f4b