mononoke/repo_client: remove direct usage of Tokio 0.1
Summary:
Like it says in the title. This removes places where we use Tokio 0.1 directly
in repo client. We use it for timeouts, so this updates us to Tokio 0.2
timeouts.

Where possible, I've made a few improvements to the code as well:

- I removed timeouts on `future::ok()` because a future that is immediately
  ready isn't going to time out.
- I updated some code to async / await where it made sense to do so to avoid
  round-tripping through futures 0.1 and 0.2 several times.

One thing that changes here is that we'll show Tokio's error on timeouts (which
says timeout has elapsed) instead of ours, which used to say just "timeout". I
think this doesn't make a big difference.

Reviewed By: StanislavGlebik

Differential Revision: D26485575

fbshipit-source-id: 8158f709bcc52d123a95df541aaeb1ec0fc9c281