Move Thrift Manifold client blobstore into its own modules
Summary:
I'm going to start shifting from Thrift to C++ client - to make reuse easier, put the Thrift client in its own module, with the Thrift-specific tests, and put shared tests in a shared module.

While in here, rename the blobstore to ManifoldBlobstore, so that it's clear that it's not just a Thrift blobstore, ready for it being a mix of C++ and Thrift.

Reviewed By: ahornby

Differential Revision: D26513599

fbshipit-source-id: 9e111583f1700eea18e83032b9ca1b9c8babc83e