mononoke: log identities as norm vector
Summary:
In EdenAPI this is logged as a vector (and in all our other services), but in
Mononoke Server we log it as a string. Let's fix this up. This is worth doing
now since right now we end up logging to 2 columns with the same name and a
different type.

Reviewed By: ahornby

Differential Revision: D26542737

fbshipit-source-id: 2f12c9e475061b1c21c71bade99b83cc070006e8