mononoke: shard walk by sql tier and shard number
Summary:
Scrubbing a repo is highly concurrent as its mostly IO bound.  As such we can end up waiting on sql connection pool for connections where it allows less than scheduled_max connections.

This change makes bounded_traversal_unique calls from the walker aware of the database tier and shard a Node may connect to, so that execution can be limited to the bounds of what the connection pool can support without waiting.

We still end up waiting for the connection, but now it's done in bounded_traversal_unique, rather than in connection pool code, and are thus a) able to process other Nodes while waiting and b) not subject to connection pool timeouts.

Differential Revision: D26524074

fbshipit-source-id: 19125388c730f5cef7e9de34b5b550efa8e6b825