test-scs-modify-bookmarks: wait for mononoke to catch up
Summary:
The `wait_for_bookmark_update` function only waits for `mononoke_scs_server` to
update to the new bookmark value.  `mononoke` might still be refreshing the
warm bookmark cache at this point, and might not have the value for `newkilo`
yet.  Flush its bookmarks cache to ensure we get an up-to-date view.

Differential Revision: D26609089

fbshipit-source-id: c7c063e4eb96303d4fbd597c4153335a3c9d4429